requirements.txt
Django==1.7.6
argparse==1.2.1
cssselect==0.9.1
flup==1.0.2
lxml==3.4.2
pymongo==2.8
pyquery==1.2.9
python-dateutil==2.4.1
pytz==2014.10
six==1.9.0
wsgiref==0.1.2
djangorestframework==3.2.3
